Allow checking API status without an instance
There's no need to have an actual HabitRPG instance for checking the API
status, particularly now we require providing login detail when creating
a HabitRPG instance.  Allow calling HabitRPG.status() as a class method
rather than an instance method.

This includes creating a static method for the actual HabitRPG API query
that doesn't require a HabitRPG instance but does require any headers
(including login details) to be provided, and a separate method that
calls the first to perform API requests that need to be authorized.