Add script for managing recurring tasks
It's pretty basic at the moment, but this script will look after
creating a new task some time after a previous task has completed.

It stores data about the tasks in YAML files in a hard-coded
subdirectory of the user's home directory; the idea being that users can
edit these as necessary, and eventually will be able to write them to
create new recurring tasks, but for the program to also be able to
update them easily.

Currently the easiest way to create a new recurring task is to use a
function available in the script; best way to access that is to import
the script in the REPL.