Add creating/updating tasks with checklists
When creating or updating a Todo or Daily, the corresponding `new` and
`update` methods now have a `checklist` argument which takes an iterable
of tuples specifying the checklist items and whether they should be
initialized as completed or not.

There's no obvious way to create the list using CheckItem objects; it
might have made sense to have a new object for about-to-be-created
checklist items, but that seemed overkill when a simple tuple would
suffice.