Handle API responses with check items missing IDs
When the API returns a response for a new task that has checklist items,
it apparently doesn't include the IDs of those checklist items.  For
classes using the ChecklistTaskMixin, override the
`create_from_api_response` method to re-fetch a task if it has checklist
items that don't have IDs, since explicitly GETting the task appears to
be sufficient to get the checklist item IDs.