Add support for deletion of recurring tasks
Catch a 404 error indicating a task has been deleted, and set it to
recur at some future time.  Determine that future time based on the
newly configured YAML values in a similar way to regular repeats; this
changes the format of the YAML file and so there is upgrade code to
handle that change.

This was written on my phone so isn't well formatted; PEP8 compliance
can come later.  It also doesn't include the facility to set up new
recurring tasks correctly.