Add parsing tags on tasks
Rather than tasks just storing the raw tag response data, parse that
data and record it using Tag objects.

Because the API will return a list of tags ID codes that includes tags
that once existed and were assigned to a task, but have now been
deleted, also add handling for deleted tasks.  This works by storing all
the tag ID codes and their corresponding Tag objects in the User object,
and if any tags don't appear on the user API response but are seen
associated with a task, they must be deleted.

This also has the nice side effect of ensuring there is only one Tag
object for a given User object and tag ID code, as customising instance
creation to return the existing object if it exists was necessary to be
able to ensure we can find the relevant Tag object again to be able to
mark it as deleted.