Add ability to remove tags from tasks
Removing a tag from a task is really a case of setting its presence as
False; there's no way to actually remove all reference to it once it's
been added.  As a result, the previous code, which assumed that
specifying the list of tags to the API would result in that list being
used, was unable to remove tags.

Instead, when updating the tag list on a task, mark all the tags that
are already on the task as not being present, then mark all the
requested tags as present, meaning any tags that aren't in the list of
tags for after the update will be marked as not present.